pychecker/requirements.txt
flask>=0.8
SQLAlchemy>=0.8.0
Flask-Login>=0.1.3
Flask-WTF>=0.8.3
PassLib>=1.6.1
alembic>=0.5.0
ghost.py>=0.1b2
PyQt>=4.10

